backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
psycopg[binary]
celery
redis
python-jose[cryptography]
passlib[bcrypt]
google-auth
google-auth-oauthlib
google-api-python-client
PyJWT
pytz